services/ingestion/requirements.txt
flask
flask-swagger-ui
requests
kafka-python
python-dotenv
prometheus-client
